redo the login feature using redirect page
